Fixing my new GUI and errors while running
Changing the background and fonts in the main tab so they would be more monotonous. Delete untested and unworking sound system embeded in Timer class. Also I was trying to fix Togles as they break the program while it runs, as of now I wasn't able to fix it entirely.