Major update for visual and logic
So. For one I played around with GUI so that the tab Widgets would fill the entire screen immideatly, instead of filling only part of the main Tk() frame. Also, for the RSA tab I changed the font and text size to match the one used in main tab. Additionaly,I played with widget placement to center the widgets in rsa tab. Additionaly, after long debuging process I fixed the problem causing immediate crash of main_tab, it was somebody's (might have been me) decision to rename the run() function in Timer to run_t() so the timer never started, resulting in a game thinking that timer has ended.

Stay tuned for more updates and debugging!