Final prelimenary fix
Gentlemen. FInaly, the bomb works as indended and currently to the best of my knowledge there are no bugs. The Toggle class was the final thing that was breaking the game. From here on, we can freely work on adding functionality to the game. Please, CONSULT and TEST BEFORE PUSHING ANYTHING.